requirements-dev.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.6
idna==2.6
olefile==0.44
Pillow==4.1.1
PyJWT==1.5.3
python-decouple==3.1
pytz==2017.2
redis==2.10.6
six==1.11.0
sorl-thumbnail==12.4a1
urllib3==1.22
django-storages==1.6.5
boto3==1.4.7
channels==1.1.8
celery==4.1.1
django-channels-presence==0.0.9
asgi-redis==1.4.3
bleach==2.1.2
markdown==2.6.10
asgiref==2.0.1
Twisted==17.9.0
whitenoise
django-allauth==0.36.0
django-tastypie==0.14.1
